mandelbrot_py/pyproject.toml
[tool.poetry]
name = "mandelbrot_py"
version = "0.1.0"
description = ""
authors = ["deepgreenAN <asami73dgreen63@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
numba = "^0.56.4"
Pillow = "^9.3.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
mypy = "^0.990"
autopep8 = "^2.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
